Quiz-Evaluation/requirements.txt
python-dotenv>=1.0.1
pydantic>=2.9.0
google-genai>=0.1.0
